Add a default config
Add a file `default_config.toml` with some basic configuration.  It is
loaded instead of the `config.toml` if the latter is not found.

The purpose of this is to have a basic, runnable config always
available, so no need to craft your own config.toml if you just want to
test something.