Access config via attributes instead of dict-like access
Pyright was complaining that the config dataclass objects don't have
getitem access.
It was actually working, because our type hints on the config object
are lying a bit.  We claim that it's an `AiTutorConfig` instance to get
better editor support but it is actually some OmegaConf thing, which
_does_ have both access options.
However, using the attribute-access makes pyright happy and is nicer
anyway.