Use a "lifespan task" to initialize the app
Calling initialization functions directly in the aitutor.py module is
bad as they get executed not only when running the application but also
in some other situations (building the frontend, database operations,
etc.).
Unfortunately, Reflex doesn't have an `on_start` event for this.
However, there are "lifespan tasks" which can be used for this.

It is not perfect, as they are running asynchronously, so the app is not
waiting for the task to be finished, but as long as the initialization
steps are fast, this should not be a problem in practice.

An open question is, if it is possible to shut the application down, in
case of some error during config loading.  With the implementation now,
an exception will be printed on the terminal but the app keeps running,
which is not ideal.


See https://github.com/orgs/reflex-dev/discussions/5320