Stop application if loading config fails
Note: Calling `sys.exit()` in the lifespan task seems to kill the
backend but not the `reflex run` process.  So the application is not
operational anymore but also not fully terminating.
This is not so great (ideally, I'd like `reflex run` to terminate with
an error code) but the best I could find for now.
The discussion
[here](https://github.com/orgs/reflex-dev/discussions/5320#discussioncomment-13281876)
suggests that there is currently no better solution.