docker: Install app dependencies in separate layer
In a first step, only mount the uv.lock and pyproject.toml files to
install the dependencies.  Then copy the whole source and install the
app.
This way, docker creates a separate layer with the dependencies only, so
we don't need to reinstall them all the time when rebuilding the images
after some code change.