Avoid use of `router.page`
`State.route.page` has been deprecated in Reflex 0.8.1.  The suggestion
in the corresponding warning message is to use `route.url` instead but I
couldn't find any information on how to replace `route.page.param`
accordingly (`route.url` seems to not be mentioned anywhere in the
docs...).
Anyway, in this case, it doesn't seem necessary to actually use
`router.page` to access the URL parameter.  Instead it can directly be
read from `self.exercise_id` in the State (as is also done in other
places of the code).  This way, we lose the default value but the page
should anyway not load if `exercise_id` is not provided, so we don't
really need a default here.  So I think this change should not cause any
issues.  At the same time, it silences the warning and simplifies the
code.