fix: Use `submit_time_stamp` to filter results without submission
The `!= []` comparison for the JSON-field `finished_conversation` does
not work like this with postgresql.
Since we anyway also have the field `submit_time_stap`, which contains
the timestamp of the last submission and is NULL if there is no
submission, simply use that instead.
That way, the filter should work fine for both Sqlite and Postgresql.