Bump reflex and reflex-local-auth
- Bump reflex to 0.8.9
- Bump reflex-local-auth to 0.4.0 (fixes a deprecation warning)

Some change in version management: To make the regular upgrade of
Reflex easier, do not pin the version in pyproject.toml but let uv take
care of it.  That is, the version is still pinned in uv.lock but since
pyproject.toml is not restricting it, we can easily bump to the latest
version by running `uv lock --upgrade` without the need to manually
edit files.