Add initial_value to computed vars that depend on the database
Before this change, building the application (e.g. via `reflex export`)
did already require a properly set up database as it is accessed in some
computed state vars.
This is now avoided by setting an `initial_value` for those variables,
which is then used at compile time.

Note: Due to https://github.com/reflex-dev/reflex/issues/5767, it is
currently not enough to just add the initial value to the function A
that accesses the database but it also needs to be done for follow-up
functions B that internally use A (or C if it uses B, etc.).