Rename TEACHER to TUTOR
Rename the "TEACHER" role to "TUTOR" which fits better to the intended
use.
Also add an alembic migration to take care of this.  Note that while
it's possible to *add* entries to an enum in postgresql, it is not
possible to *remove* them, so the TEACHER will remain, it simply won't
be used anymore (same for TUTOR when downgrading).  This should not
really be a problem, as we don't have any important production system
running which would need to be migrated.