fix mixup of localuser.id and userinfo.id
The tables localuser and userinfo are linked via userinfo.user_id (which
is a foreign key to localuser.id).  Most importantly, localuser.id and
userinfo.id are **not guaranteed to match!**

In this case, we had a mixup, where localuser.id (given by
`self.user_id`) was wrongly used instead of userinfo.id, which resulted
in an attempt to load/save the conversation of a different user.

Fixed by using userinfo.id instead of `self.user_id` everywhere in
ChatState.  For convenience, the ID is stored in a state variable that
is set by `on_load` (mostly to avoid having to put the "not None"
asserts everywhere).