Use href instead of on_click for links
The links at the bottom of the login/registration forms (pointing to the
other form) were not always working (see #233).
Instead of of an `on_click` event, use `href` to make it a regular link.
This works more reliably (at least I could not reproduce the issue
anymore) and is probably anyway better practice for ordinary links like
this (no need for javascript here).