Use component state for password_input
Use rx.ComponentState for the `password_input` component.  This way,
each input has its own toggle state (previously, all shared the same
state if there where multiple inputs on the same page) and the
additional state mixin is not needed anymore.

I also moved it from the pages/login_and_registration to the shared
components module, as it is used on multiple pages in the meantime.
Since it's the only component there for now, I just put it into the
`__init__.py`, but we can easily move it to some submodule later if
needed.