Pin postgres container version
This is the version currently used on the servers.  Pin it for now to
avoid any unpleasant surprises while updating.

For the next semester, we should probably rather use a version tag
(e.g. `postgres:18`) instead of a specific image but for now, this seems
to be the safest route for me.